Setup WebHook Bucket Notifications
A WebHook endpoint is first configured in the Minio container's environment;
this creates the "arn:minio:sqs::_:webhook" ARN, which can be used to setup
bucket notifications. A secret random token S3_WEBHOOK_SECRET is set to
provide security on the endpoint.

Bucket notifications are then setup on lossless songs bucket creation,
a flag has been added to prevent this kind of configuration in a production
environment - where we may want to configure these things manually.

The WebHook endpoint is finally created as a route in API Server, it is
currently no-op and logs the received notification JSON object.

Related to #3