Switch worker image to Alpine
The new image is one third of the original size (from 387MB to 128MB).
It also reuses the same common layers from the other images

Worker binary dependencies (shaka-packager and audiowaveform) are no longer
compiled when the transcoder-worker image is built but are instead retrieved
from a remote location and added in, allowing for faster builds and deploys.

Build scripts used for compilation are inside the 'contrib' directory.